rag/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
python-dotenv==1.0.0
pinecone-client==2.2.4
openai==1.3.7
boto3==1.34.34
google-cloud-storage==2.14.0
azure-storage-blob==12.19.0
requests==2.31.0
langchain==0.1.20
langchain-openai==0.0.5
langchain-community>=0.0.38
pydantic==2.4.2
pypdf==3.17.1
tiktoken==0.5.1 